MQTT: no floating point please

+++ include/bathos/mqtt.h
@@ -1153,7 +1153,7 @@ struct mqtt_client {
 	 *
 	 * @note This is tracked using a exponential-averaging.
 	 */
-	double typical_response_time;
+	unsigned long typical_response_time;
 
 	/**
 	 * @brief The callback that is called whenever a publish is received
